broker: convert to interthread overlay channel
Problem: the broker moves message to and from the overlay
code using functions and callbacks, which won't work if the
overlay code moves to a built-in module.

Change the interface to a message channel consisting of
back-to-back interthread:// handles which can work when
the ends of the channel are in different threads.
Messages sent by the broker to this channel go out on
the overlay.  Messages received on the overlay are
selectively sent to the broker on this channel.

This change also moves some overlay routing decisions
from the broker to the overlay code.  Having a more clearly
demarcated interface may improve maintainability of
both components.

Adapt overlay unit test.