mergify: disable temporary PR branches
Problem: Mergify is creating temporary PR branches and merging the
main branch (plus possibly other PRs). The merge commit causes the
pr-validator to fail, plus in the case of an up-to-date original PR,
the checks are redundant.

This is occurring because the maximum parallel checks default was
updated to 5 by mergify:

 https://changelog.mergify.com/changelog/maxparallelchecks-is-now-5-by-default

Explicitly set max_parallel_checks to 1 to avoid the temporary
branches. Also set batch_size to 1 in queue_rules for good measure,
since we currently never want to batch up PRs for merging.