broker: decouple state machine from overlay
Problem: the broker state machine uses overlay function calls to
manipulate and monitor the overlay network, which prevents overlay
from moving to a built-in module.

Drop all overlay function calls from the state machine.

Monitor the state machine (via streaming RPC) in the overlay.

Post events from the overlay:
- Post parent-fail on parent failure (and add a new RUN->CLEANUP
  transition)
- Post children-none or children-done in SHUTDOWN state.

Modify the shutdown warning timer so that it makes RPCs to the
overlay to find the children that are causing the delay.