flux-exec: add `--bg` flag
Problem: Processes can now be run in the background using
libsubprocess, but flux-exec does not provide a way to request this
feature.

Add a `--bg` option to flux-exec which uses `flux_rexec_bg(3)` to
launch the requested command, and exits after printing the rank and
pid of the remote process or processes.