modprobe: abort all unscheduled tasks on error
Problem: Currently, flux-modprobe exits with failure when any task
fails, but not before completing the execution of all active tasks.
This could result in cascading failures since subsequent tasks that
depend on the failing task could themselves fail in unexpected ways,
increasing the chance of unrecoverable, permanent errors (e.g. KVS
corruption)

When a task fails, do not schedule any further tasks to run.  Cancel
any started but pending tasks in the thread pool, but wait for running
tasks to complete since killing them could cause undefined behavior.
This guarantees that when a task fails, it prevents the execution of
any task that was after the failing task in the global ordering.

Fixes #7105