t: redirect both stdout/stderr in fsck tests
Problem: Most of the tests in t2816-fsck-cmd.t redirect stderr
to an output file ending in ".out".  This is because all output
in flux-fsck is sent to stderr.  In the near future, some
output may go to stdout and some to stderr.  Some additional
flexibility in the tests will be needed soon.

Redirect stdout to the file ending in ".out" and redirect stderr
to a file ending in ".err".  Update all tests accordingly.