broker: take over broker.uuid from overlay
Problem: the broker uuid is needed in the broker, but it is currently
"owned" by the overlay which may be moving to another thread.

Generate the uuid in the broker proper and set broker.uuid attribute.
Access the uuid via the broker.uuid attribute in the overlay code.
For consistency, use strlcpy(3) when copying uuids throughout the broker
and overlay code.

Update overlay unit test.