python: add `attr_set` method to Flux handle class
Problem: There is an explicit `handle.attr_get()` call in the Python
bindings but no corresponding `handle.attr_set()`.

Make a simple wrapper for `flux_attr_set(3)` in the Python Flux
handle class which not allows the function to be discoverable in the
docs. It also raises a more sensible error when the caller does
not pass a string for either of the arguments.

Fixes #7115