modprobe: support `!` prefix in needs_attrs and needs_config
Problem: The needs_attrs and needs_config arguments to the modprobe
@task decorator allow a task to be disabled if the named attribute
or config key is set, but there is no way to do the reverse: make a
task dependent on an attribute or config key _not_ being set.

Support a `!` prefix in the strings passed to needs_attrs and
needs_config to reverse the sense of the test: the task will be
disabled if the config key or attribute *is* set.

This also fixes an issue where attributes or config keys that were
set but returned a value that evaluates to false would cause a task
to be disabled if it depended on that key or attribute.

Fixes #7125