testsuite: improve test-inception.sh script
Problem: The inception test runner displays a list of job failures
at the end of the script, but most of the tests have the name `sh`,
which is unhelpful.

Add job names to all submitted jobs that are the name of the test
script.  Use a custom `flux jobs` output format so the whole name of
the test script will be be displayed.

Create .output files instead of .log files, since this matches the
suffix of files created by `make check FLUX_TEST_LOGFILES=t`.