libflux: add _new suffix to flux_set_conf()
Problem: flux_set_conf() steals the reference to its
config object argument, but isn't named in a way that would
suggest this behavior.

Lately we've been naming functions that steal references with
a _new() suffix, styled after the jansson API.  Do that here.

Update users.

There do not appear to be any users outside of flux-core.