libflux: change config-reload decode function
Problem: flux_conf_reload_decode() takes pains to relieve the
caller from having to decref the returned flux_conf_t, but most
of the time it is passed in to flux_set_conf_new(), so an incref
is necessary.

Change the function to return a non-const flux_conf_t that the
caller takes ownership of.

Rename to flux_module_config_request_decode() and relocate the
implementation to module.[ch], since it is a helper for broker
modules.

Stop using it to handle the broker config.load request in the
broker even though the payloads are identical, as that might
be confusing and it's not that helpful.

Update users.