testsuite: improve reliability of t4583-free-range-test.sh
Problem: The test `t4583-free-range-test.sh` run from
`t4000-issues-tests.t` is not reliable in ci. From the logs of a
failed run, the problem appears when the job shell is killed long
enough before the broker that the job execution system is notified
of it. This causes the job to exit with wait status 9 instead of 0.

Disconnect the broker before killing it to ensure it is unable to
report exit of the job shell. Wait for the node-failure exception before
killing processes to address any possibility of the race condition.