flux-start: don't reactivate exit timeout
Problem: the exit timer SIGKILL is sent multiple times to the same
brokers without waiting for them to respond to the last signal.

This is a logic error.  The timer watcher is started multiple times
while waiting for children.  Since flux_timer_watcher_reset() is not
called, it fires immediately.

Keep some state and only start the timer if it hasn't run yet.
One SIGKILL per broker should be sufficient.