doc: improve warning about special characters in flux-jobs(1)
Problem: The warning about the shell interpreting special characters
in flux-jobs(1) is buried in the middle of the OUTPUT FORMAT section
and may be missed by users. Also, the note is brief and doesn't
mention tcsh specifically, which is particularly problematic since
it interprets braces as special.

Add an expanded note to the beginning of the OUTPUT FORMAT section
so that it is unlikely to be missed.