python: suggest missing braces when UtilConfig format not found
Problem: When a csh or tcsh user fails to quote a format passed to one
of the utilities supporting a `--format` option, shell substitution
strips the braces and the UtilConfig class thinks the user is requested
a named format. In this case, the "No such format" error raised may
be confusing.

Enhance the error message to incdlude "Did you mean '{name}'"
to suggest that perhaps the braces have been stripped and quoting
is needed.