python: raise error for invalid input to job validator
Problem: When a jobspec alone is passed as input to the job validator,
no check is made for invalid input and an invalid ValidatorJobInfo
object is created without any other errors. This will cause validators
to fail in unpredictable ways and confuse users.

Detect missing keys in input and raise an error immediately.  In most
cases, the problem will be that the user passed the output of `flux
run --dry-run` to the validator without supplying the `--jobspec-only`
option. Hint that this may be the case in the error message.