flux-dump: internally use array instead of list
Problem: In the near future we may wish to asynchronously
load blobrefs to increase parallelization.  This can lead to
RPCs responding out of order, so using a list and appending to
it will not work for storing responses.

Convert from using a msglist to a simple array to store responses.