libflux: add flux_log_set_hostname()
Problem: flux_log(3) makes an internal call to flux_get_rank(3)
which deadlocks if the flux_t handle is not connected to a live
broker or the rank is not faked.

Add flux_log_set_hostname() which can be used to override the
internal rank call.  Once the rank is available, it can be
called with a NULL argument to restore normal operation.

Add a unit test for the new function and a couple of old ones.
Fix neighboring references to non-existent constants in flog.h.