flux-dump: support parallel dump of valref
Problem: The blobrefs within a KVS valref tree object are looked
up synchronously one by one.  This can be slow.

Support a new option --fast that will parallelize the valref lookups.
Rate cap the lookups at 1000 RPCs.

The reason this is not default is because flux-dump may have use
cases where we want the dump to be slow, such as on a running system,
and do not want it to interfere with other broker activities.

Fixes #6782