flux-dump: support parallel dump of valref
Problem: The blobrefs within a KVS valref tree object are looked
up synchronously one by one.  This can be slow.

Support a new option --maxreqs that will parallelize the valref lookups
up to the specified concurrency level.

flux-dump may have use cases where we want the dump to be slow, such as
on a running system, and do not want it to interfere with other broker
activities.  So we want a default that will give us noticeable speedup
but not overwhelm the broker.  After some experimentation, we go with a
default of 2.

Fixes #6782