broker: adjust log stderr severity/mode with --verbose
Problem: --verbose causes certain messages to be unconditionally
logged to stderr, but we are converting to a unified logging
interface that filters messages based on log severity.

When --verbose is specified, adjust logging parameters if they
were not explicitly set on the command line:

Set log-stderr-mode=local
Set log-stderr-level to
   -v:  6 (LOG_INFO)
   -vv: 7 (LOG_DEBUG)

Messages can still be conditionally logged based on the verbosity
level.  This just ensures they are seen if enabled.