libsubprocess: add optional command label
Problem: When executing subprocesses in the background, there is
no way to refer to the subprocess besides the pid, which may be
inconvenient for use cases that want to refer to commands in bulk
or start a background process in rc1 and terminate it in rc3.

Add an optional process label, set on the command object. This label
is serialized along with the command object and can then be used in
the server to look up a subprocess or otherwise refer to it by a name
as an alternative to PID.