libsubprocess: support command labels in subprocess server
Problem: Users can set command labels for use with remote processes
and the `flux_rexec*` API, but currently the subprocess server does
not do anything with these labels.

Capture command labels when set in the server and store them in
a hash. Return an error if a duplicate label is used.

Allow subprocesses to be signaled by label instead of process ID.

Return labels in the rexec.list response.