python: cli: support plugin-specific help via --help=OPTION
Problem: There is no way for cli plugins to provide extra help to
users besides the standard `help=` line in argparse output.

Add a help method to the base CLIPlugin class that prints the docstring
for the plugin to stdout. Allow plugins to override this help method
to provide extra help in a manner of the implementor's choosing.

Extend the `--help` option for all commands derived from `MiniCmd`
to take an optional argument. If an argument is provided, then call
the matching plugin's `help()` method.