libsubprocess: add FLUX_SUBPROCESS_FLAGS_WAITABLE
Problem: RFC 42 specifies a waitable flag for the exec request,
but this is not supported in libsubprocess.

Add a new FLUX_SUBPROCESS_FLAGS_WAITABLE public flag, supported only in
flux_rexec_bg(). When this flag is set, a new SUBPROCESS_REXEC_WAITABLE
flag is set in the `flags` member of the exec request.

Nothing is done with this flag yet.