libsubprocess: add FLUX_SUBPROCESS_FLAGS_WAITABLE
Problem: RFC 42 specifies a waitable flag for the exec request,
but this is not supported in libsubprocess.

Add a new rexec-only FLUX_SUBPROCESS_FLAGS_WAITABLE public flag. When
this flag is set in flux_rexec(3) or flux_rexec_bg(3), set a new
SUBPROCESS_REXEC_WAITABLE flag in the `flags` member of the exec
request.

Nothing is done with this flag yet.