libsubprocess: support waitable flag in subprocess server
Problem: The RFC 42 subprocess protocol defines a waitable flag and
wait RPC, but the subprocess server does not currently have waitable
process support.

Add support to the subprocess server for the waitable flag and
wait RPC.

Waitable processes are held in the server's processes list until the
finished status has been sent to a client or the server is shutdown.