libsubprocess: add subprocess state to process list response
Problem: Subprocesses in the server process list may either be
active or inactive now that waitable processes are supported,
but the server list response does not indicate that state.

Add a single character state to the list response. Use "R" for
running and "Z" for zombie (waiting on wait RPC).