testsuite: add option to rexec test program to wait and kill
Problem: There is no method to easily test waiting for active
subprocess that then exits in the testsuite.

Extend the rexec.py test program to add a `--wait` option to the
`kill` subcommand. This option issues a wait RPC before then going
on to kill the subprocess. This ensures the wait is started on an
active subprocess before it exits.