flux-fsck: support --job-aware
Problem: A single corrupted entry in a job directory will
effectively make all data in the job directory unusable (i.e. if
one piece of data is corrupted, other uncorrupted data may not be
usable).  The --repair option only moves the corrupted data to
the lost+found, leaving the uncorrupted data in place.  This can effect
several job related modules, that expect specific job data to always
be available.

Support a new --job-aware option to flux-fsck.  In concert with the
--repair option, if any data in a job directory is corrupted, move
all contents of the job directory to the lost+found.

Fixes #7121