connectors: avoid potential hang in ssh error path
Problem: When the ssh connector is initialized with `errp != NULL`
and `usock_client_create()` fails, the connector attempts to read all
data from the stderr file descriptor so the text can be passed back
to the caller. However, this can lead to a hang if the popen2 child
has not closed stderr yet because stderr is still in blocking mode.

Set the popen2 stderr file descriptor to non-blocking mode before
reading all data to avoid the hang.