connectors: give a hint when ssh connector fails with no stderr
Problem: A typical failure mode of the ssh connector is caused by
user shell init files that contain commands which write to stdout
in non-interactive mode. This causes the read of the initial zero
byte in the uconn protocol to fail. In this case, there is no error
emitted by ssh or the remote command (typically flux-relay(1)).

When `usock_client_create()` fails in the ssh connector, and there
is no data read from the remote command on stderr after the failure,
assume that the problem is likely related to unexpected data written
to stdout from shell init files. Write an error string that indicates
the connector failed to establish a connection, and include a hint
that the user should check their shell init files.