testsuite: test that ssh connector doesn't hang on stdout pollution
Problem: No tests check that the ssh connector does not hang when
there is extra data written to stdout, e.g. in the case of commands
writing to stdout in shell init files.

Add a test to `t1106-ssh-connector.t` that simulates this behavior.
Check that `flux_open_ex()` does not hang, and that the expected
error is returned to the caller.