python: pin slurm uri resolver to first allocated node
Problem: The slurm uri resolver uses `srun --overlap` to reinvoke
`flux uri` on the first node of the target allocation, but this is
not reliable on some versions of Slurm which round-robin across nodes
even with `srun --overlap -N1 -n1`. This can cause `flux uri` to fail
since it is likely that other nodes in the allocation do not have a
broker child on which to resolve the `FLUX_URI` environment variable.

Add `-w <first-node>` to the srun command by querying the job's nodelist
with squeue. This forces the job step to run on the first allocated
node instead of being reassigned.

Fixes #7178