testsuite: add mock squeue script to slurm uri resolver test
Problem: The slurm uri resolver now uses squeue to query the nodelist
for the target job, but a wrapper for this command is not in the
testsuite so the test will fail.

Add an squeue wrapper in t2802-uri-cmd.t. Test that squeue failure is
handled as well as empty output.