shell: rcalc: add function to normalize cores/gpus per task
Problem: In jobspec v1 with unequal distribution of tasks to nodes,
extra slots may be allocated due the requirement that there only be
one node entry in the resources section. This can lead to different
numbers of cores assigned to tasks across a job.

Add a function in the rcalc module to normalize the number of cores
or gpus per task across a job by discarding extra resources.