libsubprocess: improve comment for flux_rexec_bg()
Problem: It has been noted that the comment for `flux_rexec_bg()`
does not make it clear enough that the future can be unpacked with
`flux_rpc_get_unpack()`.

Reword the comment indicating that the payload is specifically an RPC
payload, and mention it may be unpacked with `flux_rpc_get_unpack()`.