jobtap: add flux_jobtap_set_load_sort_order(3)
Problem: When a jobtap plugin is loaded, the order in which active
jobs are introduced to the plugin via `job.new` and `job.create`
callbacks is not guaranteed.

Add flux_jobtap_set_load_sort_order(3) which, when called from
flux_plugin_init(), allows a plugin to request one of two non-default
sort orders: "state" to order by state, then jobid, or "-state" to
order by the reverse of the state then jobid. The sort order "none"
is the default and a no-op.

This allows plugins that depend on an initial order to request it
while avoiding the sort (and any changes) for plugins that do not
require a specific order.