broker: simplify response routing
Problem: the broker uses overlay functions to match the next
hop of a response message to parent or child, but it no longer
needs to do that.

If the next hop of a response message does not belong to a local
broker module, then just forward it to the overlay and let the
overlay route it.

This change revealed that the overlay code logs unnecessary
errors if it fails to route a response to an offline child.
Since overlay_uuid_is_child() returns false for an offline child,
prior to this change, responses to offline children fell through
to modhash_response_sendmsg_new(), which does not log EHOSTUNREACH
errors.  Now they (more intuitively) fall through to the overlay
which does.  Since offline peers are to be expected, the log messages
are inappropriate.  Nix them.  Also, update the overlay unit test to not
expect these log messages.