libpmi: add upmi_create_ex() with optional args
Problem: there is no way to pass arbitrary data to a upmi
plugin implementation.

Add upmi_create_ex() which is identical to upmi_create(),
except that it allows for an optional json_t *args argument.
If the argument is NULL, it is ignored.  Otherwise, its contents
are added to the "upmi.preinit" IN arguments.