python: validate host constraints against instance hostlist
Problem: Hostlist constraints set in jobspec are currently validated
only to determine if they conform to RFC 29 Hostlist format. Invalid
hosts required in a constraint currently result in an empty set of
potential resources in the scheduler and thereby an "unsatisfiable"
error, and invalid excluded hosts are silently ignored. Neither of
these are helpful to users who may have mistyped the intended hosts.

When possible, fetch the enclosing hostlist and ensure all hosts in
a constraint are contained within it, and raise an exception calling
out any invalid hostnames.

Add a function to create an on-demand, per-thread Flux handle in
the Jobspec module for use by the validation functions. Use this to
fetch the local hostlist attribute (if there is one) and cache it,
and validate any hosts in constraints using that hostlist.

Fixes #7200