libfluxutil: add conf_bootstrap_parse()
Problem: config file parsing and validating for the [bootstrap]
table is co-mingled with broker topology and overlay initialization,
making the code difficult to refactor.

Add conf_bootstrap_parse() to libfluxutil.  Take a slightly different
approach than used in the broker: generate the rank-ordered hosts
array with bind/connect URIs fully populated instead of generating
them on demand.  This results in a larger hosts array since not all
are needed, but results in a simplified interface and better separation
of concerns.  The effect on bootstrap time of an El Capitan scale
configuration seems minimal.