job-manager: improve error message from limit-job-size plugin
Problem: The error message from the limit-job-size plugin does not
include the requested resource count or the target queue. This could
leave users confused about the source of the policy.

Add the requested resource count as computed by the plugin, as well
as any queue name when generating error messages in the limit-job-size
plugin.