t/request: support check for error message
Problem: The t/request/rpc helper tool helps check if an RPC
returns a specific error number.  It would also be convenient
if it could check for an expected error message.

Support checks for a substring in the error message in addition to
the error number.  We check for a substring instead of the exact string,
since many times we do not need to check the full error message, but
rather the "core" of the error message.