docs: readme mention of contributing
Problem: the Code of Conduct (TBA), Governance, contacts, and
how to contribute (Contributing guide) need to be easy to find
at the root of the repository.
Solution: add an expected Contributing section that references
RFC 1. There is currently not CoC content there, but we will
add soon. If/when this template is good for flux-core we will
extend to other flux-framework projects.

Signed-off-by: vsoch <vsoch@users.noreply.github.com>