docker: create manifest for non-multi-arch images on tag
Problem: When docker images are pushed as part of CI after
a flux-core tag, no fluxrm/flux-core:IMAGE docker tags are
created for non-multi-arch images. Instead there is only a
fluxrm/flux-core:IMAGE-TAG. For multi-arch images, a manifest is
created at fluxrm/flux-core:IMAGE combining all the multi-arch images,
but for non-multi-arch images the manifest step is skipped.

Do not skip creation of the manifest if the manifest tag differs from
the image tag. This will create a manifest at the expected name for
images created from a tag.

Fixes #7225