job-list: stream responses from job-list
Problem: The job-list module responds to requests with a single large
array of all jobs.  This response can be extremely large, and could
lead to head-of-line blocking in the TBON.

Instead, stream job-list responses to requestors, one job at a time.
Maintain backwards compatibility by sending the entire array of all
jobs if the RPC request is not a streaming RPC.

Fixes #7190