bug: python validation multiplies string
Problem: When we use the walk helper, the resource count
assumes the inputs to be integers. If they are strings, we
get multiplied string output.
Solution: validate the resource count to be a positive
integer. This includes a test for it in a new file
anticipating wanting to add more validation tests.

Signed-off-by: vsoch <vsoch@users.noreply.github.com>