broker: add broker.shutdown-timeout attribute
Problem: there is a periodic warning timer, but no timeout on
the broker SHUTDOWN state.

This has caused problems on real systems where nodes can be
unresponsive without dropping their Flux TCP connection.

Add an optional (off by default) timeout that can be set as
a broker attribute.  It will be set for the system instance
in the systemd unit file.