systemd: set broker.shutdown-timeout=5m
Problem: shutdown hangs if a follower node gets stuck.

Add a 5m timeout on the leader broker SHUTDOWN state.
In combination with the default 1m SHUTDOWN warn timeout
which calls out nodes that are slow in shutting down, there
should be good information in the logs about why the timeout
occurred.

If timeout does occur, shutdown is still "successful" in terms of
the leader broker exit code.  It just logs an error message.

Fixes #7236