job-list: add version to list response
Problem: Newer job-list clients assume that the job-list service is always
streaming responses.  This is not the case if the client tries to communicate
with an older job-list service.

Per RFC43 update, have newer job-list responses send a job-list response
indicating the version.  This version can be used by newer clients to communicate
with both older and newer versions of the job-list protocol.

Fixes #7235