flux-job: parse version from job-list service
Problem: The job-list service will optionally respond with a version
field.  This version field can be used to determine if we are communicating
with an older or newer version of the job-list protocol.  The version field
is not currently parsed.

Parse the version field from the job-list service response.  If the version
is version 0, it indicates that future responses will not be sent from the
service and we can break out of the primary response loop.