resource: parse [resource.verify] config table
Problem: The rlist library supports a verify configuration object
initialized from a table for more granular resource verification,
but the resource module doesn't yet support it.

Parse any [resource.verify] config table and store it in a new
`verify` member of the config. Now that the config needs memory
management, add init and deinit functions for it.

Parse the table both at module startup and configuration reload.

To preserve previous behavior, if no [resource.verify] is set, then
default to then equivalent of:

 [resource.verify]
 default = "allow-extra"
 hostname = "strict"
 gpu = "ignore"

If a verify config is set, even if empty, then the "strict" default is
preserved.

Nothing is done with the config yet.