resource: drop nodrain parameter from topo_verify()
Problem: topo_verify() has no effect when the nodrain parameter is
true. In this case, the function calls rlist_verify_ex() but does
nothing with the result.

Skip topo_verify() when nodrain=true would have been set, i.e. when
the instance is a job and the resource information was obtained from
the parent's job-info service.

Drop the nodrain parameter and local variable.