job-manager: avoid in-tree perilog use-after-free
Problem: When flux_get_hostbyrank() is modified to register a new
aux item internally, there is a segfault in the perilog sharness test
when the job manager is unloaded at the end of the test.

perilog.so statically links with libflux-core.la in-tree so that the
destructor for flux_get_hostbyrank()'s aux context is located in the
perilog.so address space.  perilog.so has been unloaded by the time the
job manager's flux_t handle is called.

Fix perilog.so's in-tree linkage to ensure it doesn't get its own
copy of the destructor.

Fixes #7243.