python: update CLI plugin interface to support multiple paths
Problem: Currently CLI plugins are only loaded from a single path
(FLUX_CLI_PLUGINPATH) which overrides the default path if set.

Support a colon-delimited list for the FLUX_CLI_PLUGINPATH and load
plugins from paths provided in that variable. For debugging, add a
`--list-plugins` command-line option in the base CLI interfaces so
that users can inspect which plugins are loaded, and from where.

Fixes #6980