t: clean up CLI plugins test and associated extra tests
Problem: when t2717-python-cli-plugins.t was written, only a single
path was supported for the FLUX_CLI_PLUGINPATH environment variable,
which led to some redundant symlinks in the t/cli-plugins directory
to be able to test failure cases.

Update the test to use colon-delimited paths and eliminate the
redundant paths.