broker: set hostlist attr in bootstrap_finalize()
Problem: code for setting the hostlist broker attribute
is duplicated in boot_pmi and boot_config.

Add a new function, bootstrap_finalize(), which sets the hostlist
and calls upmi_finalize().  This function is called by the overlay
after second phase of bootstrap (endpoint exchange) has occurred.
When overlay is moved to a built-in module, it will be invoked via RPC.

In addition, make bootstrap_finalize() idempotent and call it from
the state machine after the built-in modules are initialized.
This leaves the door open for skipping overlay initialization on
a singleton instance where there is no overlay network.